+++ internal/providerfwk/provider.go
@@ -191,6 +191,7 @@ func (p *junosProvider) Resources(ctx context.Context) []func() resource.Resourc
 		newSecurityIkeGatewayResource,
 		newSecurityIkePolicyResource,
 		newSecurityIkeProposalResource,
+		newSecurityIpsecPolicyResource,
 		newSecurityIpsecProposalResource,
 		newSecurityPolicyResource,
 		newSecurityPolicyTunnelPairPolicyResource,