+++ internal/providerfwk/resource_security_ipsec_policy.go
@@ -1,0 +1,564 @@
+package providerfwk
+
+import (
+	"context"
+	"fmt"
+	"strings"
+
+	"github.com/jeremmfr/terraform-provider-junos/internal/junos"
+
+	"github.com/hashicorp/terraform-plugin-framework-validators/listvalidator"
+	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
+	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+	balt "github.com/jeremmfr/go-utils/basicalter"
+)
+
+// Ensure the implementation satisfies the expected interfaces.
+var (
+	_ resource.Resource                   = &securityIpsecPolicy{}
+	_ resource.ResourceWithConfigure      = &securityIpsecPolicy{}
+	_ resource.ResourceWithValidateConfig = &securityIpsecPolicy{}
+	_ resource.ResourceWithImportState    = &securityIpsecPolicy{}
+)
+
+type securityIpsecPolicy struct {
+	client *junos.Client
+}
+
+func newSecurityIpsecPolicyResource() resource.Resource {
+	return &securityIpsecPolicy{}
+}
+
+func (rsc *securityIpsecPolicy) typeName() string {
+	return providerName + "_security_ipsec_policy"
+}
+
+func (rsc *securityIpsecPolicy) junosName() string {
+	return "security ipsec policy"
+}
+
+func (rsc *securityIpsecPolicy) Metadata(
+	_ context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse,
+) {
+	resp.TypeName = rsc.typeName()
+}
+
+func (rsc *securityIpsecPolicy) Configure(
+	ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse,
+) {
+	// Prevent panic if the provider has not been configured.
+	if req.ProviderData == nil {
+		return
+	}
+	client, ok := req.ProviderData.(*junos.Client)
+	if !ok {
+		unexpectedResourceConfigureType(ctx, req, resp)
+
+		return
+	}
+	rsc.client = client
+}
+
+func (rsc *securityIpsecPolicy) Schema(
+	_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse,
+) {
+	resp.Schema = schema.Schema{
+		Description: "Provides a " + rsc.junosName() + ".",
+		Attributes: map[string]schema.Attribute{
+			"id": schema.StringAttribute{
+				Computed:    true,
+				Description: "An identifier for the resource with format `<name>`.",
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.UseStateForUnknown(),
+				},
+			},
+			"name": schema.StringAttribute{
+				Required:    true,
+				Description: "The name of IPSec policy.",
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+				Validators: []validator.String{
+					stringvalidator.LengthBetween(1, 32),
+					newStringFormatValidator(defaultFormat),
+				},
+			},
+			"description": schema.StringAttribute{
+				Optional:    true,
+				Description: "Text description of IPSec policy.",
+				Validators: []validator.String{
+					stringvalidator.LengthBetween(1, 900),
+					newStringDoubleQuoteExclusionValidator(),
+				},
+			},
+			"pfs_keys": schema.StringAttribute{
+				Optional:    true,
+				Description: "Diffie-Hellman Group.",
+				Validators: []validator.String{
+					stringvalidator.LengthAtLeast(1),
+					newStringSpaceExclusionValidator(),
+				},
+			},
+			"proposals": schema.ListAttribute{
+				ElementType: types.StringType,
+				Optional:    true,
+				Description: "IPSec proposals list.",
+				Validators: []validator.List{
+					listvalidator.SizeAtLeast(1),
+					listvalidator.ValueStringsAre(
+						stringvalidator.LengthBetween(1, 32),
+						newStringFormatValidator(defaultFormat),
+					),
+				},
+			},
+			"proposal_set": schema.StringAttribute{
+				Optional:    true,
+				Description: "Types of default IPSEC proposal-set.",
+				Validators: []validator.String{
+					stringvalidator.OneOf(
+						"basic",
+						"compatible",
+						"prime-128",
+						"prime-256",
+						"standard",
+						"suiteb-gcm-128",
+						"suiteb-gcm-256",
+					),
+				},
+			},
+		},
+	}
+}
+
+type securityIpsecPolicyData struct {
+	ID          types.String   `tfsdk:"id"`
+	Name        types.String   `tfsdk:"name"`
+	Description types.String   `tfsdk:"description"`
+	PfsKeys     types.String   `tfsdk:"pfs_keys"`
+	Proposals   []types.String `tfsdk:"proposals"`
+	ProposalSet types.String   `tfsdk:"proposal_set"`
+}
+
+type securityIpsecPolicyConfig struct {
+	ID          types.String `tfsdk:"id"`
+	Name        types.String `tfsdk:"name"`
+	Description types.String `tfsdk:"description"`
+	PfsKeys     types.String `tfsdk:"pfs_keys"`
+	Proposals   types.List   `tfsdk:"proposals"`
+	ProposalSet types.String `tfsdk:"proposal_set"`
+}
+
+func (rsc *securityIpsecPolicy) ValidateConfig(
+	ctx context.Context, req resource.ValidateConfigRequest, resp *resource.ValidateConfigResponse,
+) {
+	var config securityIpsecPolicyConfig
+	resp.Diagnostics.Append(req.Config.Get(ctx, &config)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	if !config.Proposals.IsNull() && !config.ProposalSet.IsNull() {
+		resp.Diagnostics.AddAttributeError(
+			path.Root("proposals"),
+			"Conflict Configuration Error",
+			"only one of proposals or proposal_set must be specified",
+		)
+	}
+}
+
+func (rsc *securityIpsecPolicy) Create(
+	ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse,
+) {
+	var plan securityIpsecPolicyData
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &plan)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+	if plan.Name.ValueString() == "" {
+		resp.Diagnostics.AddAttributeError(
+			path.Root("name"),
+			"Empty Name",
+			"could not create "+rsc.junosName()+" with empty name",
+		)
+
+		return
+	}
+
+	if rsc.client.FakeCreateSetFile() {
+		junSess := rsc.client.NewSessionWithoutNetconf(ctx)
+
+		if errPath, err := plan.set(ctx, junSess); err != nil {
+			if !errPath.Equal(path.Empty()) {
+				resp.Diagnostics.AddAttributeError(errPath, "Config Set Error", err.Error())
+			} else {
+				resp.Diagnostics.AddError("Config Set Error", err.Error())
+			}
+
+			return
+		}
+
+		plan.fillID()
+		resp.Diagnostics.Append(resp.State.Set(ctx, plan)...)
+
+		return
+	}
+
+	junSess, err := rsc.client.StartNewSession(ctx)
+	if err != nil {
+		resp.Diagnostics.AddError("Start Session Error", err.Error())
+
+		return
+	}
+	defer junSess.Close()
+	if !junSess.CheckCompatibilitySecurity() {
+		resp.Diagnostics.AddError(
+			"Compatibility Error",
+			fmt.Sprintf(rsc.junosName()+" not compatible "+
+				"with Junos device %q", junSess.SystemInformation.HardwareModel),
+		)
+
+		return
+	}
+	if err := junSess.ConfigLock(ctx); err != nil {
+		resp.Diagnostics.AddError("Config Lock Error", err.Error())
+
+		return
+	}
+	policyExists, err := checkSecurityIpsecPolicyExists(ctx, plan.Name.ValueString(), junSess)
+	if err != nil {
+		resp.Diagnostics.Append(diagWarns("Config Clear Warning", junSess.ConfigClear())...)
+		resp.Diagnostics.AddError("Pre Check Error", err.Error())
+
+		return
+	}
+	if policyExists {
+		resp.Diagnostics.Append(diagWarns("Config Clear Warning", junSess.ConfigClear())...)
+		resp.Diagnostics.AddError(
+			"Duplicate Configuration Error",
+			fmt.Sprintf(rsc.junosName()+" %q already exists", plan.Name.ValueString()),
+		)
+
+		return
+	}
+
+	if errPath, err := plan.set(ctx, junSess); err != nil {
+		resp.Diagnostics.Append(diagWarns("Config Clear Warning", junSess.ConfigClear())...)
+		if !errPath.Equal(path.Empty()) {
+			resp.Diagnostics.AddAttributeError(errPath, "Config Set Error", err.Error())
+		} else {
+			resp.Diagnostics.AddError("Config Set Error", err.Error())
+		}
+
+		return
+	}
+	warns, err := junSess.CommitConf("create resource " + rsc.typeName())
+	resp.Diagnostics.Append(diagWarns("Config Commit Warning", warns)...)
+	if err != nil {
+		resp.Diagnostics.Append(diagWarns("Config Clear Warning", junSess.ConfigClear())...)
+		resp.Diagnostics.AddError("Config Commit Error", err.Error())
+
+		return
+	}
+
+	policyExists, err = checkSecurityIpsecPolicyExists(ctx, plan.Name.ValueString(), junSess)
+	if err != nil {
+		resp.Diagnostics.Append(diagWarns("Config Clear Warning", junSess.ConfigClear())...)
+		resp.Diagnostics.AddError("Post Check Error", err.Error())
+
+		return
+	}
+	if !policyExists {
+		resp.Diagnostics.AddError(
+			"Not Found Error",
+			fmt.Sprintf(rsc.junosName()+" %q not exists after commit "+
+				"=> check your config", plan.Name.ValueString()),
+		)
+
+		return
+	}
+
+	plan.fillID()
+	resp.Diagnostics.Append(resp.State.Set(ctx, plan)...)
+}
+
+func (rsc *securityIpsecPolicy) Read(
+	ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse,
+) {
+	var state, data securityIpsecPolicyData
+	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+	junSess, err := rsc.client.StartNewSession(ctx)
+	if err != nil {
+		resp.Diagnostics.AddError("Start Session Error", err.Error())
+
+		return
+	}
+	defer junSess.Close()
+
+	junos.MutexLock()
+	err = data.read(ctx, state.Name.ValueString(), junSess)
+	junos.MutexUnlock()
+	if err != nil {
+		resp.Diagnostics.AddError("Config Read Error", err.Error())
+
+		return
+	}
+	if data.ID.ValueString() == "" {
+		resp.State.RemoveResource(ctx)
+
+		return
+	}
+
+	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
+}
+
+func (rsc *securityIpsecPolicy) Update(
+	ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse,
+) {
+	var plan, state securityIpsecPolicyData
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &plan)...)
+	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	if rsc.client.FakeUpdateAlso() {
+		junSess := rsc.client.NewSessionWithoutNetconf(ctx)
+
+		if err := state.del(ctx, junSess); err != nil {
+			resp.Diagnostics.AddError("Config Del Error", err.Error())
+
+			return
+		}
+		if errPath, err := plan.set(ctx, junSess); err != nil {
+			if !errPath.Equal(path.Empty()) {
+				resp.Diagnostics.AddAttributeError(errPath, "Config Set Error", err.Error())
+			} else {
+				resp.Diagnostics.AddError("Config Set Error", err.Error())
+			}
+
+			return
+		}
+
+		resp.Diagnostics.Append(resp.State.Set(ctx, plan)...)
+
+		return
+	}
+
+	junSess, err := rsc.client.StartNewSession(ctx)
+	if err != nil {
+		resp.Diagnostics.AddError("Start Session Error", err.Error())
+
+		return
+	}
+	defer junSess.Close()
+	if err := junSess.ConfigLock(ctx); err != nil {
+		resp.Diagnostics.AddError("Config Lock Error", err.Error())
+
+		return
+	}
+
+	if err := state.del(ctx, junSess); err != nil {
+		resp.Diagnostics.Append(diagWarns("Config Clear Warning", junSess.ConfigClear())...)
+		resp.Diagnostics.AddError("Config Del Error", err.Error())
+
+		return
+	}
+	if errPath, err := plan.set(ctx, junSess); err != nil {
+		resp.Diagnostics.Append(diagWarns("Config Clear Warning", junSess.ConfigClear())...)
+		if !errPath.Equal(path.Empty()) {
+			resp.Diagnostics.AddAttributeError(errPath, "Config Set Error", err.Error())
+		} else {
+			resp.Diagnostics.AddError("Config Set Error", err.Error())
+		}
+
+		return
+	}
+	warns, err := junSess.CommitConf("update resource " + rsc.typeName())
+	resp.Diagnostics.Append(diagWarns("Config Commit Warning", warns)...)
+	if err != nil {
+		resp.Diagnostics.Append(diagWarns("Config Clear Warning", junSess.ConfigClear())...)
+		resp.Diagnostics.AddError("Config Commit Error", err.Error())
+
+		return
+	}
+
+	resp.Diagnostics.Append(resp.State.Set(ctx, plan)...)
+}
+
+func (rsc *securityIpsecPolicy) Delete(
+	ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse,
+) {
+	var state securityIpsecPolicyData
+	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	if rsc.client.FakeDeleteAlso() {
+		junSess := rsc.client.NewSessionWithoutNetconf(ctx)
+
+		if err := state.del(ctx, junSess); err != nil {
+			resp.Diagnostics.AddError("Config Del Error", err.Error())
+
+			return
+		}
+
+		return
+	}
+
+	junSess, err := rsc.client.StartNewSession(ctx)
+	if err != nil {
+		resp.Diagnostics.AddError("Start Session Error", err.Error())
+
+		return
+	}
+	defer junSess.Close()
+	if err := junSess.ConfigLock(ctx); err != nil {
+		resp.Diagnostics.AddError("Config Lock Error", err.Error())
+
+		return
+	}
+
+	if err := state.del(ctx, junSess); err != nil {
+		resp.Diagnostics.Append(diagWarns("Config Clear Warning", junSess.ConfigClear())...)
+		resp.Diagnostics.AddError("Config Del Error", err.Error())
+
+		return
+	}
+	warns, err := junSess.CommitConf("delete resource " + rsc.typeName())
+	resp.Diagnostics.Append(diagWarns("Config Commit Warning", warns)...)
+	if err != nil {
+		resp.Diagnostics.Append(diagWarns("Config Clear Warning", junSess.ConfigClear())...)
+		resp.Diagnostics.AddError("Config Commit Error", err.Error())
+
+		return
+	}
+}
+
+func (rsc *securityIpsecPolicy) ImportState(
+	ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse,
+) {
+	junSess, err := rsc.client.StartNewSession(ctx)
+	if err != nil {
+		resp.Diagnostics.AddError("Start Session Error", err.Error())
+
+		return
+	}
+	defer junSess.Close()
+
+	policyExists, err := checkSecurityIpsecPolicyExists(ctx, req.ID, junSess)
+	if err != nil {
+		resp.Diagnostics.AddError("Pre Check Error", err.Error())
+
+		return
+	}
+	if !policyExists {
+		resp.Diagnostics.AddError(
+			"Not Found Error",
+			fmt.Sprintf("don't find "+rsc.junosName()+" with id %q "+
+				"(id must be <name>)", req.ID),
+		)
+
+		return
+	}
+
+	var data securityIpsecPolicyData
+	err = data.read(ctx, req.ID, junSess)
+	if err != nil {
+		resp.Diagnostics.AddError("Config Read Error", err.Error())
+
+		return
+	}
+
+	if !data.ID.IsNull() {
+		resp.Diagnostics.Append(resp.State.Set(ctx, data)...)
+	}
+}
+
+func checkSecurityIpsecPolicyExists(_ context.Context, name string, junSess *junos.Session) (bool, error) {
+	showConfig, err := junSess.Command(junos.CmdShowConfig +
+		"security ipsec policy " + name + junos.PipeDisplaySet)
+	if err != nil {
+		return false, err
+	}
+	if showConfig == junos.EmptyW {
+		return false, nil
+	}
+
+	return true, nil
+}
+
+func (rscData *securityIpsecPolicyData) fillID() {
+	rscData.ID = types.StringValue(rscData.Name.ValueString())
+}
+
+func (rscData *securityIpsecPolicyData) set(_ context.Context, junSess *junos.Session) (path.Path, error) {
+	configSet := make([]string, 0)
+
+	setPrefix := "set security ipsec policy " + rscData.Name.ValueString() + " "
+	if v := rscData.Description.ValueString(); v != "" {
+		configSet = append(configSet, setPrefix+"description \""+v+"\"")
+	}
+	if v := rscData.PfsKeys.ValueString(); v != "" {
+		configSet = append(configSet, setPrefix+"perfect-forward-secrecy keys "+v)
+	}
+	for _, v := range rscData.Proposals {
+		configSet = append(configSet, setPrefix+"proposals "+v.ValueString())
+	}
+	if v := rscData.ProposalSet.ValueString(); v != "" {
+		configSet = append(configSet, setPrefix+"proposal-set "+v)
+	}
+
+	return path.Empty(), junSess.ConfigSet(configSet)
+}
+
+func (rscData *securityIpsecPolicyData) read(_ context.Context, name string, junSess *junos.Session,
+) (err error) {
+	showConfig, err := junSess.Command(junos.CmdShowConfig +
+		"security ipsec policy " + name + junos.PipeDisplaySetRelative)
+	if err != nil {
+		return err
+	}
+	if showConfig != junos.EmptyW {
+		rscData.Name = types.StringValue(name)
+		rscData.fillID()
+		for _, item := range strings.Split(showConfig, "\n") {
+			if strings.Contains(item, junos.XMLStartTagConfigOut) {
+				continue
+			}
+			if strings.Contains(item, junos.XMLEndTagConfigOut) {
+				break
+			}
+			itemTrim := strings.TrimPrefix(item, junos.SetLS)
+			switch {
+			case balt.CutPrefixInString(&itemTrim, "description "):
+				rscData.Description = types.StringValue(strings.Trim(itemTrim, "\""))
+			case balt.CutPrefixInString(&itemTrim, "perfect-forward-secrecy keys "):
+				rscData.PfsKeys = types.StringValue(itemTrim)
+			case balt.CutPrefixInString(&itemTrim, "proposals "):
+				rscData.Proposals = append(rscData.Proposals, types.StringValue(itemTrim))
+			case balt.CutPrefixInString(&itemTrim, "proposal-set "):
+				rscData.ProposalSet = types.StringValue(itemTrim)
+			}
+		}
+	}
+
+	return nil
+}
+
+func (rscData *securityIpsecPolicyData) del(_ context.Context, junSess *junos.Session) error {
+	configSet := []string{
+		"delete security ipsec policy " + rscData.Name.ValueString(),
+	}
+
+	return junSess.ConfigSet(configSet)
+}