+++ CHANGELOG.md
@@ -26,6 +26,9 @@ ENHANCEMENTS:
 * **resource/junos_security_ike_proposal**:
   * resource now use new [terraform-plugin-framework](github.com/hashicorp/terraform-plugin-framework) (optional string attributes doesn't accept *empty* value)
   * add `description` argument
+* **resource/junos_security_ipsec_policy**:
+  * resource now use new [terraform-plugin-framework](github.com/hashicorp/terraform-plugin-framework) (optional string attributes doesn't accept *empty* value)
+  * add `description` argument
 * **resource/junos_security_ipsec_proposal**:
   * resource now use new [terraform-plugin-framework](github.com/hashicorp/terraform-plugin-framework) (optional string attributes doesn't accept *empty* value)
   * add `description` argument